requirements.txt
Flask==3.0.3
google-genai==0.3.0
google-cloud-storage==2.14.0
werkzeug==3.0.1
